cros_ec_lpc_io_bytes_mec can return -ve codes for error.

+++ drivers/platform/chrome/cros_ec_lpc_mec.c
@@ -118,9 +118,9 @@ int cros_ec_lpc_mec_in_range(unsigned int offset, unsigned int length)
  *
  * Return: 8-bit checksum of all bytes read / written
  */
-u8 cros_ec_lpc_io_bytes_mec(enum cros_ec_lpc_mec_io_type io_type,
-			    unsigned int offset, unsigned int length,
-			    u8 *buf)
+int cros_ec_lpc_io_bytes_mec(enum cros_ec_lpc_mec_io_type io_type,
+			     unsigned int offset, unsigned int length,
+			     u8 *buf)
 {
 	int i = 0;
 	int io_addr;
